dure: move Win32 file replacement into the session store PAL leaf
The `MoveFileExW` call sat inline in the portable session store, so on the
Linux runners its mutants compiled into nothing and could never be killed. It
now lives in the platform leaf alongside the rest of the Windows PAL.

`canonicalize` gained a test pinning its result to the real resolved path,
which was previously indistinguishable from an empty one off Windows.

Copilot-Session: 073848cb-c418-46e4-a97c-0714f4b3c953
